LayoutTests/svg/dom/SVGViewSpec-invalid-ref-crash.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="container"><svg xmlns="http://www.w3.org/2000/svg" width="200" height="200"/></div>
<script>
    description("Confirm that exceptions are thrown, or empty string returned, when an SVGViewSpec is used after its corresponding element has been removed.");

    successfullyParsed = false;

    var svgView = document.getElementById("container").childNodes[0].currentView;
    document.getElementById("container").removeChild(document.getElementById("container").childNodes[0]);

    gc();

    shouldThrow("viewPreserveAspectRatio = svgView.preserveAspectRatio;");
    shouldBeNull("viewTransform = svgView.transform;");
    shouldThrow("viewViewTarget = svgView.viewTarget;");
    shouldThrow("viewViewBox = svgView.viewBox;");

    successfullyParsed = true;

</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
